+++ src/notification/notification.service.ts
@@ -10,6 +10,15 @@ export class NotificationService {
         private readonly transactionRepo: Repository<Transaction>,
     ){}
 
+    async createTransaction(performerId: number, receiverId: number, description: string){
+        let result = await this.transactionRepo.save({
+            performerId: performerId,
+            receiverId: receiverId,
+            description: description
+        })
+        return result
+    }
+
     // get all transaction for receiverId
     async getNotificationsFor(receiverId: number){
         let results = await this.transactionRepo.find({
